split things up
