+++ pi_service/src/main.rs
@@ -2,46 +2,19 @@
 extern crate crossbeam_channel as channel;
 extern crate wiringpi;
 
-use std::collections::HashMap;
+mod model;
+mod pi_receiver;
+mod pins;
+
+use model::PiOutput;
+use pins::*;
 use std::thread;
 use std::time::Duration;
-
-enum PiOutput {
-    Led { pin: u16, value: i32 },
-    Blink { pin: u16, millis: u32 },
-}
-
-const RED_GPIO: u16 = 12;
-const GREEN_GPIO: u16 = 16;
-const BLUE_GPIO: u16 = 20;
-
-fn pi_consumer(output_r: channel::Receiver<PiOutput>) {
-    // Setup wiringPi in GPIO mode (with original BCM numbering order)
-    let pi = wiringpi::setup_gpio();
-    let pins = {
-        let mut p = HashMap::new();
-
-        // track some pins
-        p.insert(RED_GPIO, pi.soft_pwm_pin(RED_GPIO)); // red
-        p.insert(GREEN_GPIO, pi.soft_pwm_pin(GREEN_GPIO)); // green
-        p.insert(BLUE_GPIO, pi.soft_pwm_pin(BLUE_GPIO)); // blue
-        p
-    };
-    loop {
-        match output_r.recv() {
-            Some(PiOutput::Led { pin, value }) => if let Some(p) = pins.get(&pin) {
-                p.pwm_write(value);
-            },
-            Some(PiOutput::Blink { pin: _, millis: _ }) => {}
-            None => {}
-        }
-    }
-}
 
 fn main() {
     let (output_s, output_r) = channel::bounded(1);
 
-    thread::spawn(move || pi_consumer(output_r));
+    thread::spawn(move || pi_receiver::run(output_r));
 
     // clear
     output_s.send(PiOutput::Led {
