+++ pi_service/src/model.rs
@@ -1,0 +1,4 @@
+pub enum PiOutput {
+    Led { pin: u16, value: i32 },
+    Blink { pin: u16, millis: u32 },
+}
