+++ pi_service/src/pins.rs
@@ -1,0 +1,3 @@
+pub const RED_GPIO: u16 = 12;
+pub const GREEN_GPIO: u16 = 16;
+pub const BLUE_GPIO: u16 = 20;
